Time: 32 ms (22.33%), Space: 100.6 MB (16.12%) - LeetHub

+++ 1435-xor-queries-of-a-subarray/1435-xor-queries-of-a-subarray.cpp
@@ -80,7 +80,7 @@ public:
 
         for (vector<int> val : queries)
         {
-            int L = val[0], R = val[1];
+            short L = val[0], R = val[1];
             if ((L-1) < 0)
                 result.emplace_back(prefixXor[R]);
             else
